+++ NEWS.md
@@ -1,3 +1,4 @@
 # trajr 1.0.1
 
 * Fixed: `plot.TrajSpeedIntervals` was not passing additional arguments (`...`) to `plot`.
+* Added: functions `TrajDuration`, `TrajMeanVelocity`.